Add users
The (very first draft of the) account system is here!

Create an account with a username and password, additionally including an email
address (optional). You can then sign in with that username and password.
Uploads now *require* you to be signed in, and carts are related to your user
account.

Passwords are stored in the database salted, with a 32-character salt appended.
The hashing function used here is BLAKE2s.