Add favorites
Favorites are here! The uploader of a cart cannot favorite it, but any other user can. The favorite button appears grey when an item hasn't been favorited, and red when it has been. Hovering over the button switches the color, indicating that clicking the button will change whether or not you have favorited the cart.